requirements.txt
mysql-connector==2.2.9
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
shiboken6==6.5.0
